consider fuzz functions as tests
Treat the case where even though the function has the Test prefix
it starts with a lowercase letter after Test, in this case it is
not a test.

Remove the stderr parameter from runtests because, apparently, when
we pass the -json flag to "go test" no data goes to standard error
output.

Create a benchmark because later I will modify the use of packages.Load
to see if it improves performance.