rewrite the tests to use the result.txt file
You can find a specification of the tests in the file test.txt.

Make the thereAreTests operation part of the tests operation, because
the tests operation cannot determine when a package has no tests. This
happens because of the --cover option. If the package has no test files
then the --cover option prevents the command from reporting it.

Handle the case where the test has a "build-fail" action. One case were
this happens is when a test function has a wrong signature.

Show only an error of one operation. For example, if the tests operation
returns an error then show only it. If multiple operations return errors
then show only the error of the first in the order that we show the
messages. A example of a case of error is when the build fail in the tests
operation. When this happens then, probably, a error will also happen in
the staticcheck operation. Show only one to not show the same error two times.